Modified player controller script
Modified the provide player controller script to the point where it's more of how I would have it. Also reworked logic to facilitate jumping in the air (resets the vertical speed for more clean jumps) as well as adding a configurable multiple jumps functionality.
Also modified the player gravity to give a less floaty feeling for player character